+++ Changelog.md
@@ -1,3 +1,10 @@
+19/06/2018
+
+NovusCodeLibrary4.Utils
+
+* New method bool IsDateEmpty(DateTime aDate) in class DateTimeUtils 
+
+
 14/11/2017
 
 * Start of the Changelog
@@ -5,8 +12,8 @@
 NovusCodeLibrary4.SimpleTemplate
 
 * Updated TagName without <%xxx%> save in HashTable for key
-* New function public string CleanTagName(string aRawTagEx)
-* New function public void UpdateTagValue(string aTagName, string aTagValue )
+* New method public string CleanTagName(string aRawTagEx)
+* New method public void UpdateTagValue(string aTagName, string aTagValue )
 * New property IgnoreBlankTags - If TagValue is blank re-insert <%TagName%>